Move OBSERVABLE_TESTS and DISPATCH_TESTS to proper tests
